Format index.html, adjust image sizes
index.html was workable but the link appeared on one line. Added
paragraph formatting to create separation and look a bit more clean.
The default sizes of the images used on the recipe pages were very large.
Set the width to 400 so that scrolling is not needed to get past images.