Add title to homepage, change image source
The title was left as the boilerplate's default "Document" so it was
changed to more accurately reflect the page's contents. The images were
were absolute references to webpages. An images directory was created,
images were downloaded to place in the directory, and the references on
the pages were changed to reference the saved directory images instead.
It makes the code look much cleaner and it is easier to edit without the
scrollbar in VSCode. I also had the thought it may be considered better
or load faster, but don't know if that is actually true.